Add files via upload
The Ramayana movie directed by Nitesh Tiwari will purportedly showcase Ranbir Kapoor as Lord Ram, Sai Pallavi as Sita, and Yash as Raavan. The legendary tale is slated to debut on the silver screen in the latter part of 2025. Nitesh is poised to commence production on his cinematic adaptation of the Hindu epic in February/March 2024, according to a recent report from Pinkvilla. Yash is set to join the Ramayana team in July after completing a significant portion of his upcoming film, Toxic.